#48: Rocks will lock Sitecore.Kernel.dll when connected
Rocks no longer locks Sitecore.Kernel to obtain the .NET Framework version.

This check was meant to prevent Rocks from automatically updating a website with assemblies which use a higher Framework version than the website.

Since this obtaining the framework version automatilcally is no longer possible, the framework version can be specified manually in the Connection Properties in the DotNetFrameworkVersion field. The default version is v10.0.0.